+++ CHANGELOG.md
@@ -6,4 +6,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### -Added
+
+- Documentation Sampel Code
+- Project Structure
+- Include file (amb.asm)
+- copyright.c